Move validator to base app
