+++ apps/tunes/models.py
@@ -1,18 +1,7 @@
 from django.db import models
-from django.core.exceptions import ValidationError
-from django.utils.translation import gettext_lazy as _
 
 from base.models import BaseModel
-
-
-def validate_decimal_value(value):
-    """Validate `value` is between 0 <= value <= 1"""
-    if value < 0 or value > 1:
-        raise ValidationError(
-            _('{} must be between 0 and 1'.format(value)),
-            params={'value': value}
-
-    )
+from base.validators import validate_decimal_value
 
 
 class Emotion(BaseModel):
